refactor: Move transaction query on history
